move message types to types/message.type.ts, replace time display with Moment format and delete time utility function
