+++ package.json
@@ -19,6 +19,7 @@
     "expo-image": "~2.3.2",
     "expo-status-bar": "~2.2.3",
     "jsdoc": "^4.0.3",
+    "moment": "^2.30.1",
     "react": "19.0.0",
     "react-dom": "19.0.0",
     "react-native": "0.79.5",
